+++ src/fhir-core/data-types/primitive/XhtmlType.ts
@@ -22,8 +22,9 @@
  */
 
 import { Extension, PrimitiveType } from '@src/fhir-core/base-models/core-fhir-models';
-import { fhirXhtml, fhirXhtmlSchema, parseFhirPrimitiveData } from './primitive-types';
+import { isDefined } from '@src/fhir-core/utility/type-guards';
 import { FhirError } from '@src/fhir-core/errors/FhirError';
+import { fhirXhtml, fhirXhtmlSchema, parseFhirPrimitiveData } from './primitive-types';
 
 /**
  * Xhtml Class
@@ -103,7 +104,7 @@ export class XhtmlType extends PrimitiveType<fhirXhtml> {
   }
 
   private assignValue(value: fhirXhtml | undefined): void {
-    if (value !== undefined) {
+    if (isDefined<fhirXhtml | undefined>(value)) {
       super.setValue(parseFhirPrimitiveData(value, fhirXhtmlSchema, this.typeErrorMessage()));
     } else {
       super.setValue(undefined);
@@ -111,7 +112,7 @@ export class XhtmlType extends PrimitiveType<fhirXhtml> {
   }
 
   private assignExtension(extension: Extension[] | undefined): void {
-    if (extension !== undefined) {
+    if (isDefined<Extension[] | undefined>(extension)) {
       throw new FhirError(this.EXTENSION_ERROR);
     }
     super.setExtension(undefined);